+++ src/game/server/tf/tf_obj.h
@@ -360,6 +360,6 @@ private:
 	CNetworkVar( bool, m_bServerOverridePlacement );
 };
 
-extern short g_sModelIndexFireball;		// holds the index for the fireball
+extern int g_sModelIndexFireball;		// holds the index for the fireball
 
 #endif // TF_OBJ_H
